backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.6.0
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
ruff>=0.8.0
mypy>=1.13.0
